progress?

+++ interpreter/src/main/java/org/enso/interpreter/node/expression/operator/AddOperatorNode.java
@@ -8,13 +8,13 @@ import org.enso.interpreter.node.expression.BinaryOperatorNode;
 public abstract class AddOperatorNode extends BinaryOperatorNode {
 
   @Specialization
-  protected long add(int left, int right) { return left + right; }
+  protected long add(int left, int right) { return (long) left + (long) right; }
 
   @Specialization
-  protected long add(long left, int right) { return left + right; }
+  protected long add(long left, int right) { return left + (long) right; }
 
   @Specialization
-  protected long add(int left, long right) { return left + right; }
+  protected long add(int left, long right) { return (long) left + right; }
 
   @Specialization
   protected long add(long left, long right) {
